Add files via upload
So, this is a website created by me, MommaTopaz as Ms. Blue Sky.  I am also Ms. Blue Sky on codepen.  This is a multi-page layout.  I pulled the photos from pexels.  I hope they don't mind.  This was my first major multi-page.  I must look at it within the next few days and make sure that the files are correct.  I've been monkeying around with the font colors as you can see on the main page (the page with the blog).  What I did here overall was create a site that has picture and menu links to the other pages and in doing so, this Theme2 can be either a one-page site or a multi-page site.  All I have to do to change to a single page site is trade the link codes out.  Now I've notated a LOT on the html and it would be a bit confusing to an outsider.  What I did is code the main page and then the additional pages and they should be in order of appearances as I am a stickler for detail.  I have not cleaned up the code too much yet.  This is just an example of a multi-page site that I designed.  Thank you for looking.